Update instructions.md
Hello, I believe there was a mistake in line 38 of the instructions file. Step 3 of the exercise originally stated, "If the User Is an Eagle," but that doesn't make sense in the context. Therefore, I believed the correct context was, "If the User Is Touching an Eagle," so I added the word, "touching." If I'm mistaken, I apologize hehe. I appreciate your support of the exercise community.